fix crud comments

+++ src/repository/comments.py
@@ -30,6 +30,10 @@ async def create_comment(content: str, user: str, photos_id: int, db: AsyncSessi
         await db.rollback()
         raise e
 
+async def get_comment(id: int, db: AsyncSession):   
+    comment = await db.get(Comment, id)
+    if comment:
+        return comment
 
 async def update_comment(text: str, id: int, db: AsyncSession):
     """
@@ -45,19 +49,21 @@ async def update_comment(text: str, id: int, db: AsyncSession):
     :rtype: Comment
     """
     
-    try:
-        comment = await db.get(Comment, id)
-        comment.text = text
-        comment.update_status = True
-        await db.commit()
-        await db.refresh(comment)
-        return comment
-    except Exception as e:
-        await db.rollback()
-        raise e
+    comment = await db.get(Comment, id)
+    if comment:
+        try:
+            comment.text = text
+            comment.update_status = True
+            await db.commit()
+            await db.refresh(comment)
+            return comment
+        except Exception as e:
+            await db.rollback()
+            raise e
+    return None
 
 
-async def delete_comment(comment_id: int, db: AsyncSession):
+async def delete_comment(id: int, db: AsyncSession):
     """
     Delete a comment
 
@@ -69,16 +75,18 @@ async def delete_comment(comment_id: int, db: AsyncSession):
     :rtype: Comment
     """
     
-    try:
-        comment = await db.get(Comment, comment_id)
-        await db.delete(comment)
-        await db.commit()
-        return comment
-    except Exception as e:
-        await db.rollback()
-        raise e
+    comment = await db.get(Comment, id)
+    if comment:
+        try:
+            await db.delete(comment)
+            await db.commit()
+            return comment
+        except Exception as e:
+            await db.rollback()
+            raise e
 
-async def get_comments(
+
+async def get_photo_comments(
     offset: int, 
     limit: int, 
     photo_id: int, 
@@ -100,5 +108,29 @@ async def get_comments(
     sq = select(Comment).filter(Comment.photo_id == photo_id).offset(offset).limit(limit)
     comments = await db.execute(sq)
     result = comments.scalars().all()
-    print(result)
+    return result
+
+
+async def get_user_comments(
+    offset: int, 
+    limit: int, 
+    user_id: int, 
+    db: AsyncSession):
+    """
+    Review comments by photo
+
+    :param limit: limit of comments
+    :type: int
+    :param offset: offset of comments
+    :type offset: int
+    :param photos_id: The ID of the photo for which to retrieve ratings.
+    :type photos_id: int
+    :param db: The database session.
+    :type db: AsyncSession
+    :return: A list of comment objects.
+    :rtype: List[Comment]
+    """
+    sq = select(Comment).filter(Comment.user_id == user_id).offset(offset).limit(limit)
+    comments = await db.execute(sq)
+    result = comments.scalars().all()
     return result
